syntax: Move config types to a separate module
